move ByteEncoder from substrate_metadata to polkadart_scale_codec
