+++ packages/polkadart_scale_codec/lib/src/core/encoder.dart
@@ -200,7 +200,6 @@ abstract class ScaleCodecEncoder {
   }
 }
 
-//TODO: Refactor [ScaleCodecEncoder] and [HexEncoder]
 /// `HexEncoder` to write the bytes and parallely write the result as Hex.
 class HexEncoder extends ScaleCodecEncoder {
   String _hex = '0x';
@@ -219,5 +218,37 @@ class HexEncoder extends ScaleCodecEncoder {
   /// Return current hex data
   String toHex() {
     return _hex;
+  }
+}
+
+/// `ByteEncoder` to process and write the bytes
+class ByteEncoder extends ScaleCodecEncoder {
+  Uint8List _data = Uint8List(128);
+  int _pos = 0;
+
+  void _alloc(int size) {
+    if (_data.length - _pos < size) {
+      Uint8List newData = Uint8List(max(size, _data.length) * 2);
+      newData.setAll(0, _data);
+      _data = newData;
+    }
+  }
+
+  @override
+  void write(int byte) {
+    _alloc(1);
+    _data[_pos] = byte;
+    _pos += 1;
+  }
+
+  @override
+  void bytes(List<int> b) {
+    _alloc(b.length);
+    _data.setAll(_pos, b);
+    _pos += b.length;
+  }
+
+  Uint8List toBytes() {
+    return _data.sublist(0, _pos);
   }
 }
